[refactor][schema/item] Do not change content. Only re-organize files
