+++ schema/item/item.go
@@ -1,17 +1,103 @@
 package item
 
 import (
+	"encoding/json"
+	"fmt"
+
 	"github.com/hashicorp/go-version"
+	"github.com/progrhyme/binq/internal/erron"
 	"github.com/progrhyme/binq/internal/logs"
 )
 
-type ReviseMode int
+// Item wraps itemProps which corresponds to JSON structure of item data
+type Item struct {
+	*itemProps
+}
 
-const (
-	ReviseModeNatural ReviseMode = iota + 1
-	ReviseModeLatest
-	ReviseModeOld
-)
+func DecodeItemJSON(b []byte) (item *Item, err error) {
+	var i itemProps
+	if _err := json.Unmarshal(b, &i); _err != nil {
+		return item, erron.Errorwf(_err, "Failed to unmarshal JSON: %s", b)
+	}
+	item = &Item{itemProps: &i}
+	return item, err
+}
+
+func (i *Item) String() string {
+	return fmt.Sprintf("%+v", *i.itemProps)
+}
+
+func (i *Item) Print(pretty bool) (b []byte, err error) {
+	var _err error
+	if pretty {
+		b, _err = json.MarshalIndent(i, "", "  ")
+	} else {
+		b, _err = json.Marshal(i)
+	}
+	if _err != nil {
+		return b, erron.Errorwf(_err, "Failed to marshal JSON: %s", i)
+	}
+	return b, nil
+}
+
+func (i *Item) GetLatestURL(param ItemURLParam) (url string, err error) {
+	rev := i.GetLatest()
+	if rev == nil {
+		return "", nil
+	}
+	return rev.GetURL(param)
+}
+
+func (i *Item) GetLatest() (rev *ItemRevision) {
+	latest := i.Latest
+	if latest.Version == "" {
+		return nil
+	}
+
+	found := i.GetRevision(latest.Version)
+	if found != nil {
+		return found
+	}
+
+	return &ItemRevision{
+		Version:      latest.Version,
+		URLFormat:    i.Meta.URLFormat,
+		Replacements: i.Meta.Replacements,
+		Extension:    i.Meta.Extension,
+	}
+}
+
+func (i *Item) GetRevision(version string) (rev *ItemRevision) {
+	tmp := &ItemRevision{
+		Version:      version,
+		URLFormat:    i.Meta.URLFormat,
+		Replacements: i.Meta.Replacements,
+		Extension:    i.Meta.Extension,
+	}
+
+	found := false
+	for _, ver := range i.Versions {
+		if ver.Version == version {
+			found = true
+			tmp.Checksums = ver.Checksums
+			if ver.URLFormat != "" {
+				tmp.URLFormat = ver.URLFormat
+			}
+			if ver.Replacements != nil {
+				tmp.Replacements = ver.Replacements
+			}
+			if ver.Extension != nil {
+				tmp.Extension = ver.Extension
+			}
+			break
+		}
+	}
+	if found {
+		return tmp
+	}
+
+	return nil
+}
 
 func (i *Item) AddOrUpdateRevision(rev *ItemRevision, mode ReviseMode) {
 	var replaced bool
